+++ src/utils/systool.py
@@ -90,3 +90,11 @@ def wait_until_memory_available(
         )
         time.sleep(REFRESH_SECONDS)
     raise Exception("No Memory available")
+
+
+def human_readable_size(size, decimal_places=2):
+    for unit in ["B", "KiB", "MiB", "GiB", "TiB", "PiB"]:
+        if size < 1024.0 or unit == "PiB":
+            break
+        size /= 1024.0
+    return f"{size:.{decimal_places}f} {unit}"